# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ConsumerApp API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/message": {
            "get": {
                "tags": [
                    "Message"
                ],
                "description": "Получение списка сообщений",
                "responses": {
                    "200": {
                        "description": "Список сообщений найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Массив найденных сообщений",
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Message"
                                                    }
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Message"
                ],
                "description": "Добавить сообщение",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Сообщение добавлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Message"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/message/{message}": {
            "get": {
                "tags": [
                    "Message"
                ],
                "description": "Получение сообщение",
                "responses": {
                    "200": {
                        "description": "Сообщение найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Message"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Message"
                ],
                "description": "Отредактировать сообщение",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Сообщение отредактировано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Message"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Message"
                ],
                "description": "Получение списка сообщений",
                "responses": {
                    "200": {
                        "description": "Сообщение удалено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "null"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MessageRequest": {
                "properties": {
                    "user_id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer"
                    },
                    "message": {
                        "description": "Сообщение",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Message": {
                "properties": {
                    "id": {
                        "description": "Идентификатор сообщения",
                        "type": "integer"
                    },
                    "message": {
                        "description": "ФИО клиента",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "send_time": {
                        "description": "Время отправки",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "description": "Идентификатор пользователя",
                        "type": "integer"
                    },
                    "email": {
                        "description": "Почта пользователя",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Когда пользователь был зарегистрирован",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}